Various checks for required attributes and better error messages added. Submitted by:	Nico Seessle <nico@seessle.de>
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@268002 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/Rename.java
@@ -102,6 +102,14 @@ public class Rename extends Task {
      * thrown, if the rename operation fails.
      */
     public void execute() throws BuildException {
+        if (dest == null) {
+            throw new BuildException("dest attribute is required", location);
+        }
+
+        if (src == null) {
+            throw new BuildException("src attribute is required", location);
+        }
+
         if (replace && dest.exists()) {
             if (!dest.delete()) {
                 throw new BuildException("Unable to remove existing file " +
